# services/gateway-origination-svc/openapi.yaml
openapi: 3.0.3
info:
  title: gateway-origination-svc
  version: 0.1.0
  description: |
    Gateway de borde + Identidad/KYC + Saga de originación (Inciso I).
    Puerto: 3331. Rol D2 — NeoLend Financial Corp.
servers:
  - url: http://localhost:3331/v1
    description: Desarrollo local
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ProblemDetails:
      type: object
      properties:
        type: { type: string }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        correlationId: { type: string }
...
    LoanApplication:
      type: object
      properties:
        applicationId: { type: string, format: uuid }
        correlationId: { type: string }
        applicantId: { type: string, format: uuid }
        requestedAmount: { type: number }
        currency: { type: string }
        termMonths: { type: integer }
        status:
          type: string
          enum: [SUBMITTED, SCORING, DECIDING, APPROVED, MANUAL, REJECTED, FAILED]
        timeline:
          type: array
          items:
            type: object
            properties:
              step: { type: string }
              status: { type: string }
              occurredAt: { type: string }
              detail: { type: string }
        scoreId: { type: string }
        score: { type: integer }
        riskBand: { type: string }
        decisionId: { type: string }
        creditId: { type: string }
        approvedAmount: { type: number }
        interestRate: { type: number }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
security:
  - bearerAuth: []
paths:
...
  /loan-applications:
    post:
      summary: Inicia solicitud de crédito (Saga)
      description: |
        Retorna 202 inmediatamente. La saga coordina scoring→decisión en background.
        El cliente hace polling a pollUrl para obtener el estado.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [applicantId, requestedAmount, termMonths]
              properties:
                applicantId: { type: string, format: uuid }
                requestedAmount: { type: number, minimum: 1 }
                currency: { type: string, default: USD }
                termMonths: { type: integer, minimum: 1 }
      responses:
        '202':
          description: Solicitud en proceso
          content:
            application/json:
              schema:
                type: object
                properties:
                  applicationId: { type: string, format: uuid }
                  status: { type: string }
                  pollUrl: { type: string }
  /loan-applications/{id}:
    get:
      summary: Estado + progreso (timeline) de la solicitud
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Estado y timeline de la solicitud
          content:
            application/json:
              schema: { $ref: '#/components/schemas/LoanApplication' }
        '404':
          description: No encontrada
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ProblemDetails' }
...